Move non-nested classes and interfaces to their own files
