+++ src/Microsoft.AspNet.Server.Kestrel/Http/Connection.cs
@@ -2,42 +2,11 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Threading.Tasks;
 using Microsoft.AspNet.Server.Kestrel.Networking;
 using System.Diagnostics;
 
 namespace Microsoft.AspNet.Server.Kestrel.Http
 {
-    public class ConnectionContext : ListenerContext
-    {
-        public ConnectionContext()
-        {
-        }
-
-        public ConnectionContext(ListenerContext context) : base(context)
-        {
-        }
-
-        public ConnectionContext(ConnectionContext context) : base(context)
-        {
-            SocketInput = context.SocketInput;
-            SocketOutput = context.SocketOutput;
-            ConnectionControl = context.ConnectionControl;
-        }
-
-        public SocketInput SocketInput { get; set; }
-        public ISocketOutput SocketOutput { get; set; }
-
-        public IConnectionControl ConnectionControl { get; set; }
-    }
-
-    public interface IConnectionControl
-    {
-        void Pause();
-        void Resume();
-        void End(ProduceEndType endType);
-    }
-
     public class Connection : ConnectionContext, IConnectionControl
     {
         private static readonly Action<UvStreamHandle, int, Exception, object> _readCallback = ReadCallback;
