+++ src/Microsoft.AspNet.Server.Kestrel/Http/FrameContext.cs
@@ -1,0 +1,17 @@
+namespace Microsoft.AspNet.Server.Kestrel.Http
+{
+    public class FrameContext : ConnectionContext
+    {
+        public FrameContext()
+        {
+
+        }
+
+        public FrameContext(ConnectionContext context) : base(context)
+        {
+
+        }
+
+        public IFrameControl FrameControl { get; set; }
+    }
+}
