+++ src/Microsoft.AspNet.Server.Kestrel/Http/IConnectionControl.cs
@@ -1,0 +1,9 @@
+namespace Microsoft.AspNet.Server.Kestrel.Http
+{
+    public interface IConnectionControl
+    {
+        void Pause();
+        void Resume();
+        void End(ProduceEndType endType);
+    }
+}
