+++ src/Microsoft.AspNet.Server.Kestrel/Http/ISocketOutput.cs
@@ -1,0 +1,12 @@
+using System;
+
+namespace Microsoft.AspNet.Server.Kestrel.Http
+{
+    /// <summary>
+    ///   Operations performed for buffered socket output
+    /// </summary>
+    public interface ISocketOutput
+    {
+        void Write(ArraySegment<byte> buffer, Action<Exception, object> callback, object state);
+    }
+}
