+++ src/Microsoft.AspNet.Server.Kestrel/Http/Listener.cs
@@ -9,24 +9,6 @@ using System.Threading.Tasks;
 
 namespace Microsoft.AspNet.Server.Kestrel.Http
 {
-    public class ListenerContext
-    {
-        public ListenerContext() { }
-
-        public ListenerContext(ListenerContext context)
-        {
-            Thread = context.Thread;
-            Application = context.Application;
-            Memory = context.Memory;
-        }
-
-        public KestrelThread Thread { get; set; }
-
-        public Func<Frame, Task> Application { get; set; }
-
-        public IMemoryPool Memory { get; set; }
-    }
-
     /// <summary>
     /// Summary description for Accept
     /// </summary>
