+++ src/Microsoft.AspNet.Server.Kestrel/Http/ListenerContext.cs
@@ -1,0 +1,23 @@
+using System;
+using System.Threading.Tasks;
+
+namespace Microsoft.AspNet.Server.Kestrel.Http
+{
+    public class ListenerContext
+    {
+        public ListenerContext() { }
+
+        public ListenerContext(ListenerContext context)
+        {
+            Thread = context.Thread;
+            Application = context.Application;
+            Memory = context.Memory;
+        }
+
+        public KestrelThread Thread { get; set; }
+
+        public Func<Frame, Task> Application { get; set; }
+
+        public IMemoryPool Memory { get; set; }
+    }
+}
