+++ src/Microsoft.AspNet.Server.Kestrel/Http/ProduceEndType.cs
@@ -1,0 +1,9 @@
+namespace Microsoft.AspNet.Server.Kestrel.Http
+{
+    public enum ProduceEndType
+    {
+        SocketShutdownSend,
+        SocketDisconnect,
+        ConnectionKeepAlive,
+    }
+}
