+++ src/Microsoft.AspNet.Server.Kestrel/Networking/Libuv.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Reflection;
 using System.Runtime.InteropServices;
-using System.Text;
 
 namespace Microsoft.AspNet.Server.Kestrel.Networking
 {
