+++ src/Microsoft.AspNet.Server.Kestrel/Networking/PlatformApis.cs
@@ -2,9 +2,7 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Reflection;
 using System.Runtime.InteropServices;
-using System.Text;
 
 namespace Microsoft.AspNet.Server.Kestrel.Networking
 {
